+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Change log
 ### 6.9.2 (10/3/2018 - )
-* httpClient: tweak request/response to make it more low level style, as we have high level api client
+* httpClient: tweak request/response to make it more low level style
 
 ### 6.9.1 (10/3/2018)
 * httpClient: fix bug on shouldRetry